huwei_0806 发布系统代码

+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/filter/ProductModuleSearchContext.java
@@ -55,15 +55,15 @@ public class ProductModuleSearchContext extends QueryWrapperContext<ProductModul
             this.getSearchCond().eq("type", n_type_eq);
         }
     }
-	private BigInteger n_root_eq;//[产品]
-	public void setN_root_eq(BigInteger n_root_eq) {
+	private Long n_root_eq;//[产品]
+	public void setN_root_eq(Long n_root_eq) {
         this.n_root_eq = n_root_eq;
         if(!ObjectUtils.isEmpty(this.n_root_eq)){
             this.getSearchCond().eq("root", n_root_eq);
         }
     }
-	private BigInteger n_parent_eq;//[id]
-	public void setN_parent_eq(BigInteger n_parent_eq) {
+	private Long n_parent_eq;//[id]
+	public void setN_parent_eq(Long n_parent_eq) {
         this.n_parent_eq = n_parent_eq;
         if(!ObjectUtils.isEmpty(this.n_parent_eq)){
             this.getSearchCond().eq("parent", n_parent_eq);
