+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/ibiz/filter/TaskTeamSearchContext.java
@@ -27,8 +27,8 @@ import cn.ibizlab.pms.core.ibiz.domain.TaskTeam;
 @Data
 public class TaskTeamSearchContext extends QueryWrapperContext<TaskTeam> {
 
-	private BigInteger n_root_eq;//[编号]
-	public void setN_root_eq(BigInteger n_root_eq) {
+	private Long n_root_eq;//[编号]
+	public void setN_root_eq(Long n_root_eq) {
         this.n_root_eq = n_root_eq;
         if(!ObjectUtils.isEmpty(this.n_root_eq)){
             this.getSearchCond().eq("root", n_root_eq);
