+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/domain/Burn.java
@@ -27,6 +27,7 @@ import cn.ibizlab.pms.util.annotation.Audit;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.baomidou.mybatisplus.annotation.*;
 import cn.ibizlab.pms.util.domain.EntityMP;
+import com.baomidou.mybatisplus.core.toolkit.IdWorker;
 
 /**
  * 实体[burn]
@@ -82,14 +83,14 @@ public class Burn extends EntityMP implements Serializable {
     @TableField(value = "project")
     @JSONField(name = "project")
     @JsonProperty("project")
-    private BigInteger project;
+    private Long project;
     /**
      * 任务
      */
     @TableField(value = "task")
     @JSONField(name = "task")
     @JsonProperty("task")
-    private BigInteger task;
+    private Long task;
 
     /**
      * 
@@ -154,7 +155,7 @@ public class Burn extends EntityMP implements Serializable {
     /**
      * 设置 [所属项目]
      */
-    public void setProject(BigInteger project){
+    public void setProject(Long project){
         this.project = project ;
         this.modify("project",project);
     }
@@ -162,7 +163,7 @@ public class Burn extends EntityMP implements Serializable {
     /**
      * 设置 [任务]
      */
-    public void setTask(BigInteger task){
+    public void setTask(Long task){
         this.task = task ;
         this.modify("task",task);
     }
