+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/domain/ProjectProduct.java
@@ -27,6 +27,7 @@ import cn.ibizlab.pms.util.annotation.Audit;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.baomidou.mybatisplus.annotation.*;
 import cn.ibizlab.pms.util.domain.EntityMP;
+import com.baomidou.mybatisplus.core.toolkit.IdWorker;
 
 /**
  * 实体[项目产品]
@@ -74,28 +75,28 @@ public class ProjectProduct extends EntityMP implements Serializable {
     @TableField(value = "product")
     @JSONField(name = "product")
     @JsonProperty("product")
-    private BigInteger product;
+    private Long product;
     /**
      * 产品计划
      */
     @TableField(value = "plan")
     @JSONField(name = "plan")
     @JsonProperty("plan")
-    private BigInteger plan;
+    private Long plan;
     /**
      * 平台/分支
      */
     @TableField(value = "branch")
     @JSONField(name = "branch")
     @JsonProperty("branch")
-    private BigInteger branch;
+    private Long branch;
     /**
      * 项目
      */
     @TableField(value = "project")
     @JSONField(name = "project")
     @JsonProperty("project")
-    private BigInteger project;
+    private Long project;
 
     /**
      * 
@@ -134,7 +135,7 @@ public class ProjectProduct extends EntityMP implements Serializable {
     /**
      * 设置 [产品]
      */
-    public void setProduct(BigInteger product){
+    public void setProduct(Long product){
         this.product = product ;
         this.modify("product",product);
     }
@@ -142,7 +143,7 @@ public class ProjectProduct extends EntityMP implements Serializable {
     /**
      * 设置 [产品计划]
      */
-    public void setPlan(BigInteger plan){
+    public void setPlan(Long plan){
         this.plan = plan ;
         this.modify("plan",plan);
     }
@@ -150,7 +151,7 @@ public class ProjectProduct extends EntityMP implements Serializable {
     /**
      * 设置 [平台/分支]
      */
-    public void setBranch(BigInteger branch){
+    public void setBranch(Long branch){
         this.branch = branch ;
         this.modify("branch",branch);
     }
@@ -158,7 +159,7 @@ public class ProjectProduct extends EntityMP implements Serializable {
     /**
      * 设置 [项目]
      */
-    public void setProject(BigInteger project){
+    public void setProject(Long project){
         this.project = project ;
         this.modify("project",project);
     }
