+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/filter/ProjectSearchContext.java
@@ -41,15 +41,15 @@ public class ProjectSearchContext extends QueryWrapperContext<Project> {
             this.getSearchCond().eq("pm", n_pm_eq);
         }
     }
-	private BigInteger n_id_eq;//[项目编号]
-	public void setN_id_eq(BigInteger n_id_eq) {
+	private Long n_id_eq;//[项目编号]
+	public void setN_id_eq(Long n_id_eq) {
         this.n_id_eq = n_id_eq;
         if(!ObjectUtils.isEmpty(this.n_id_eq)){
             this.getSearchCond().eq("id", n_id_eq);
         }
     }
-	private BigInteger n_id_noteq;//[项目编号]
-	public void setN_id_noteq(BigInteger n_id_noteq) {
+	private Long n_id_noteq;//[项目编号]
+	public void setN_id_noteq(Long n_id_noteq) {
         this.n_id_noteq = n_id_noteq;
         if(!ObjectUtils.isEmpty(this.n_id_noteq)){
             this.getSearchCond().ne("id", n_id_noteq);
@@ -139,8 +139,8 @@ public class ProjectSearchContext extends QueryWrapperContext<Project> {
             this.getSearchCond().like("parentname", n_parentname_like);
         }
     }
-	private BigInteger n_parent_eq;//[父项目]
-	public void setN_parent_eq(BigInteger n_parent_eq) {
+	private Long n_parent_eq;//[父项目]
+	public void setN_parent_eq(Long n_parent_eq) {
         this.n_parent_eq = n_parent_eq;
         if(!ObjectUtils.isEmpty(this.n_parent_eq)){
             this.getSearchCond().eq("parent", n_parent_eq);
