+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/filter/ReleaseSearchContext.java
@@ -62,22 +62,22 @@ public class ReleaseSearchContext extends QueryWrapperContext<Release> {
             this.getSearchCond().like("buildname", n_buildname_like);
         }
     }
-	private BigInteger n_product_eq;//[产品]
-	public void setN_product_eq(BigInteger n_product_eq) {
+	private Long n_product_eq;//[产品]
+	public void setN_product_eq(Long n_product_eq) {
         this.n_product_eq = n_product_eq;
         if(!ObjectUtils.isEmpty(this.n_product_eq)){
             this.getSearchCond().eq("product", n_product_eq);
         }
     }
-	private BigInteger n_build_eq;//[版本]
-	public void setN_build_eq(BigInteger n_build_eq) {
+	private Long n_build_eq;//[版本]
+	public void setN_build_eq(Long n_build_eq) {
         this.n_build_eq = n_build_eq;
         if(!ObjectUtils.isEmpty(this.n_build_eq)){
             this.getSearchCond().eq("build", n_build_eq);
         }
     }
-	private BigInteger n_branch_eq;//[平台/分支]
-	public void setN_branch_eq(BigInteger n_branch_eq) {
+	private Long n_branch_eq;//[平台/分支]
+	public void setN_branch_eq(Long n_branch_eq) {
         this.n_branch_eq = n_branch_eq;
         if(!ObjectUtils.isEmpty(this.n_branch_eq)){
             this.getSearchCond().eq("branch", n_branch_eq);
