+++ ibizpms-core/src/main/java/cn/ibizlab/pms/core/zentao/filter/StoryStageSearchContext.java
@@ -27,15 +27,15 @@ import cn.ibizlab.pms.core.zentao.domain.StoryStage;
 @Data
 public class StoryStageSearchContext extends QueryWrapperContext<StoryStage> {
 
-	private BigInteger n_story_eq;//[需求]
-	public void setN_story_eq(BigInteger n_story_eq) {
+	private Long n_story_eq;//[需求]
+	public void setN_story_eq(Long n_story_eq) {
         this.n_story_eq = n_story_eq;
         if(!ObjectUtils.isEmpty(this.n_story_eq)){
             this.getSearchCond().eq("story", n_story_eq);
         }
     }
-	private BigInteger n_branch_eq;//[平台/分支]
-	public void setN_branch_eq(BigInteger n_branch_eq) {
+	private Long n_branch_eq;//[平台/分支]
+	public void setN_branch_eq(Long n_branch_eq) {
         this.n_branch_eq = n_branch_eq;
         if(!ObjectUtils.isEmpty(this.n_branch_eq)){
             this.getSearchCond().eq("branch", n_branch_eq);
