+++ ibizpms-provider/ibizpms-provider-webapi/src/main/java/cn/ibizlab/pms/webapi/dto/ActionDTO.java
@@ -46,7 +46,7 @@ public class ActionDTO extends DTOBase implements Serializable {
      */
     @JSONField(name = "id")
     @JsonProperty("id")
-    private BigInteger id;
+    private Long id;
 
     /**
      * 属性 [COMMENT]
@@ -111,7 +111,7 @@ public class ActionDTO extends DTOBase implements Serializable {
      */
     @JSONField(name = "project")
     @JsonProperty("project")
-    private BigInteger project;
+    private Long project;
 
     /**
      * 属性 [LASTCOMMENT]
@@ -205,7 +205,7 @@ public class ActionDTO extends DTOBase implements Serializable {
     /**
      * 设置 [PROJECT]
      */
-    public void setProject(BigInteger  project){
+    public void setProject(Long  project){
         this.project = project ;
         this.modify("project",project);
     }
