+++ ibizpms-provider/ibizpms-provider-webapi/src/main/java/cn/ibizlab/pms/webapi/dto/GroupDTO.java
@@ -46,7 +46,7 @@ public class GroupDTO extends DTOBase implements Serializable {
      */
     @JSONField(name = "id")
     @JsonProperty("id")
-    private BigInteger id;
+    private Long id;
 
     /**
      * 属性 [NAME]
